+++ src/Symfony/Bridge/Monolog/Tests/LoggerTest.php
@@ -128,7 +128,7 @@ class LoggerTest extends TestCase
     public function testInheritedClassCallGetLogsWithoutArgument()
     {
         $loggerChild = new ClassThatInheritLogger('test');
-        $this->assertNull($loggerChild->getLogs());
+        $this->assertSame([], $loggerChild->getLogs());
     }
 
     public function testInheritedClassCallCountErrorsWithoutArgument()
@@ -140,9 +140,9 @@ class LoggerTest extends TestCase
 
 class ClassThatInheritLogger extends Logger
 {
-    public function getLogs(Request $request = null)
+    public function getLogs(Request $request = null): array
     {
-        parent::getLogs($request);
+        return parent::getLogs($request);
     }
 
     public function countErrors(Request $request = null)