+++ src/Symfony/Bridge/Twig/Extension/DumpExtension.php
@@ -17,7 +17,6 @@ use Symfony\Component\VarDumper\Dumper\HtmlDumper;
 use Twig\Environment;
 use Twig\Extension\AbstractExtension;
 use Twig\Template;
-use Twig\TokenParser\TokenParserInterface;
 use Twig\TwigFunction;
 
 /**
@@ -25,7 +24,7 @@ use Twig\TwigFunction;
  *
  * @author Nicolas Grekas <p@tchwork.com>
  */
-class DumpExtension extends AbstractExtension
+final class DumpExtension extends AbstractExtension
 {
     private $cloner;
     private $dumper;
@@ -37,9 +36,9 @@ class DumpExtension extends AbstractExtension
     }
 
     /**
-     * @return TwigFunction[]
+     * {@inheritdoc}
      */
-    public function getFunctions()
+    public function getFunctions(): array
     {
         return [
             new TwigFunction('dump', [$this, 'dump'], ['is_safe' => ['html'], 'needs_context' => true, 'needs_environment' => true]),
@@ -47,19 +46,14 @@ class DumpExtension extends AbstractExtension
     }
 
     /**
-     * @return TokenParserInterface[]
+     * {@inheritdoc}
      */
-    public function getTokenParsers()
+    public function getTokenParsers(): array
     {
         return [new DumpTokenParser()];
     }
 
-    public function getName()
-    {
-        return 'dump';
-    }
-
-    public function dump(Environment $env, array $context)
+    public function dump(Environment $env, array $context): ?string
     {
         if (!$env->isDebug()) {
             return null;