+++ src/Symfony/Bridge/Twig/Extension/HttpKernelExtension.php
@@ -20,12 +20,12 @@ use Twig\TwigFunction;
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-class HttpKernelExtension extends AbstractExtension
+final class HttpKernelExtension extends AbstractExtension
 {
     /**
-     * @return TwigFunction[]
+     * {@inheritdoc}
      */
-    public function getFunctions()
+    public function getFunctions(): array
     {
         return [
             new TwigFunction('render', [HttpKernelRuntime::class, 'renderFragment'], ['is_safe' => ['html']]),
@@ -34,16 +34,8 @@ class HttpKernelExtension extends AbstractExtension
         ];
     }
 
-    public static function controller(string $controller, array $attributes = [], array $query = [])
+    public static function controller(string $controller, array $attributes = [], array $query = []): ControllerReference
     {
         return new ControllerReference($controller, $attributes, $query);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return 'http_kernel';
     }
 }