+++ src/Symfony/Bridge/Twig/Extension/HttpKernelRuntime.php
@@ -19,7 +19,7 @@ use Symfony\Component\HttpKernel\Fragment\FragmentHandler;
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-class HttpKernelRuntime
+final class HttpKernelRuntime
 {
     private $handler;
 
@@ -33,11 +33,9 @@ class HttpKernelRuntime
      *
      * @param string|ControllerReference $uri A URI as a string or a ControllerReference instance
      *
-     * @return string The fragment content
-     *
      * @see FragmentHandler::render()
      */
-    public function renderFragment($uri, array $options = [])
+    public function renderFragment($uri, array $options = []): string
     {
         $strategy = isset($options['strategy']) ? $options['strategy'] : 'inline';
         unset($options['strategy']);
@@ -50,11 +48,9 @@ class HttpKernelRuntime
      *
      * @param string|ControllerReference $uri A URI as a string or a ControllerReference instance
      *
-     * @return string The fragment content
-     *
      * @see FragmentHandler::render()
      */
-    public function renderFragmentStrategy(string $strategy, $uri, array $options = [])
+    public function renderFragmentStrategy(string $strategy, $uri, array $options = []): string
     {
         return $this->handler->render($uri, $strategy, $options);
     }