+++ src/Symfony/Bridge/Twig/Extension/LogoutUrlExtension.php
@@ -20,7 +20,7 @@ use Twig\TwigFunction;
  *
  * @author Jeremy Mikola <jmikola@gmail.com>
  */
-class LogoutUrlExtension extends AbstractExtension
+final class LogoutUrlExtension extends AbstractExtension
 {
     private $generator;
 
@@ -31,10 +31,8 @@ class LogoutUrlExtension extends AbstractExtension
 
     /**
      * {@inheritdoc}
-     *
-     * @return TwigFunction[]
      */
-    public function getFunctions()
+    public function getFunctions(): array
     {
         return [
             new TwigFunction('logout_url', [$this, 'getLogoutUrl']),
@@ -46,10 +44,8 @@ class LogoutUrlExtension extends AbstractExtension
      * Generates the relative logout URL for the firewall.
      *
      * @param string|null $key The firewall key or null to use the current firewall key
-     *
-     * @return string The relative logout URL
      */
-    public function getLogoutPath(string $key = null)
+    public function getLogoutPath(string $key = null): string
     {
         return $this->generator->getLogoutPath($key);
     }
@@ -58,19 +54,9 @@ class LogoutUrlExtension extends AbstractExtension
      * Generates the absolute logout URL for the firewall.
      *
      * @param string|null $key The firewall key or null to use the current firewall key
-     *
-     * @return string The absolute logout URL
      */
-    public function getLogoutUrl(string $key = null)
+    public function getLogoutUrl(string $key = null): string
     {
         return $this->generator->getLogoutUrl($key);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return 'logout_url';
     }
 }