+++ src/Symfony/Bridge/Twig/Extension/StopwatchExtension.php
@@ -21,7 +21,7 @@ use Twig\TokenParser\TokenParserInterface;
  *
  * @author Wouter J <wouter@wouterj.nl>
  */
-class StopwatchExtension extends AbstractExtension
+final class StopwatchExtension extends AbstractExtension
 {
     private $stopwatch;
     private $enabled;
@@ -32,7 +32,7 @@ class StopwatchExtension extends AbstractExtension
         $this->enabled = $enabled;
     }
 
-    public function getStopwatch()
+    public function getStopwatch(): Stopwatch
     {
         return $this->stopwatch;
     }
@@ -40,7 +40,7 @@ class StopwatchExtension extends AbstractExtension
     /**
      * @return TokenParserInterface[]
      */
-    public function getTokenParsers()
+    public function getTokenParsers(): array
     {
         return [
             /*
@@ -50,10 +50,5 @@ class StopwatchExtension extends AbstractExtension
              */
             new StopwatchTokenParser(null !== $this->stopwatch && $this->enabled),
         ];
-    }
-
-    public function getName()
-    {
-        return 'stopwatch';
     }
 }