+++ src/Symfony/Bridge/Twig/Extension/WebLinkExtension.php
@@ -22,7 +22,7 @@ use Twig\TwigFunction;
  *
  * @author Kévin Dunglas <dunglas@gmail.com>
  */
-class WebLinkExtension extends AbstractExtension
+final class WebLinkExtension extends AbstractExtension
 {
     private $requestStack;
 
@@ -33,10 +33,8 @@ class WebLinkExtension extends AbstractExtension
 
     /**
      * {@inheritdoc}
-     *
-     * @return TwigFunction[]
      */
-    public function getFunctions()
+    public function getFunctions(): array
     {
         return [
             new TwigFunction('link', [$this, 'link']),
@@ -56,7 +54,7 @@ class WebLinkExtension extends AbstractExtension
      *
      * @return string The relation URI
      */
-    public function link(string $uri, string $rel, array $attributes = [])
+    public function link(string $uri, string $rel, array $attributes = []): string
     {
         if (!$request = $this->requestStack->getMasterRequest()) {
             return $uri;
@@ -80,7 +78,7 @@ class WebLinkExtension extends AbstractExtension
      *
      * @return string The path of the asset
      */
-    public function preload(string $uri, array $attributes = [])
+    public function preload(string $uri, array $attributes = []): string
     {
         return $this->link($uri, 'preload', $attributes);
     }
@@ -92,7 +90,7 @@ class WebLinkExtension extends AbstractExtension
      *
      * @return string The path of the asset
      */
-    public function dnsPrefetch(string $uri, array $attributes = [])
+    public function dnsPrefetch(string $uri, array $attributes = []): string
     {
         return $this->link($uri, 'dns-prefetch', $attributes);
     }
@@ -104,7 +102,7 @@ class WebLinkExtension extends AbstractExtension
      *
      * @return string The path of the asset
      */
-    public function preconnect(string $uri, array $attributes = [])
+    public function preconnect(string $uri, array $attributes = []): string
     {
         return $this->link($uri, 'preconnect', $attributes);
     }
@@ -116,7 +114,7 @@ class WebLinkExtension extends AbstractExtension
      *
      * @return string The path of the asset
      */
-    public function prefetch(string $uri, array $attributes = [])
+    public function prefetch(string $uri, array $attributes = []): string
     {
         return $this->link($uri, 'prefetch', $attributes);
     }
@@ -128,7 +126,7 @@ class WebLinkExtension extends AbstractExtension
      *
      * @return string The path of the asset
      */
-    public function prerender(string $uri, array $attributes = [])
+    public function prerender(string $uri, array $attributes = []): string
     {
         return $this->link($uri, 'prerender', $attributes);
     }