+++ src/Symfony/Bridge/Twig/Extension/YamlExtension.php
@@ -12,7 +12,6 @@
 namespace Symfony\Bridge\Twig\Extension;
 
 use Symfony\Component\Yaml\Dumper as YamlDumper;
-use Symfony\Component\Yaml\Yaml;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
 
@@ -21,14 +20,12 @@ use Twig\TwigFilter;
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-class YamlExtension extends AbstractExtension
+final class YamlExtension extends AbstractExtension
 {
     /**
      * {@inheritdoc}
-     *
-     * @return TwigFilter[]
      */
-    public function getFilters()
+    public function getFilters(): array
     {
         return [
             new TwigFilter('yaml_encode', [$this, 'encode']),
@@ -36,7 +33,7 @@ class YamlExtension extends AbstractExtension
         ];
     }
 
-    public function encode($input, int $inline = 0, int $dumpObjects = 0)
+    public function encode($input, int $inline = 0, int $dumpObjects = 0): string
     {
         static $dumper;
 
@@ -51,7 +48,7 @@ class YamlExtension extends AbstractExtension
         return $dumper->dump($input, $inline, 0, false, $dumpObjects);
     }
 
-    public function dump($value, int $inline = 0, int $dumpObjects = 0)
+    public function dump($value, int $inline = 0, int $dumpObjects = 0): string
     {
         if (\is_resource($value)) {
             return '%Resource%';
@@ -62,13 +59,5 @@ class YamlExtension extends AbstractExtension
         }
 
         return $this->encode($value, $inline, $dumpObjects);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return 'yaml';
     }
 }