+++ src/Symfony/Bridge/Twig/TokenParser/DumpTokenParser.php
@@ -27,14 +27,12 @@ use Twig\TokenParser\AbstractTokenParser;
  *
  * @author Julien Galenski <julien.galenski@gmail.com>
  */
-class DumpTokenParser extends AbstractTokenParser
+final class DumpTokenParser extends AbstractTokenParser
 {
     /**
      * {@inheritdoc}
-     *
-     * @return Node
      */
-    public function parse(Token $token)
+    public function parse(Token $token): Node
     {
         $values = null;
         if (!$this->parser->getStream()->test(Token::BLOCK_END_TYPE)) {
@@ -47,10 +45,8 @@ class DumpTokenParser extends AbstractTokenParser
 
     /**
      * {@inheritdoc}
-     *
-     * @return string
      */
-    public function getTag()
+    public function getTag(): string
     {
         return 'dump';
     }