+++ src/Symfony/Bridge/Twig/TokenParser/StopwatchTokenParser.php
@@ -22,7 +22,7 @@ use Twig\TokenParser\AbstractTokenParser;
  *
  * @author Wouter J <wouter@wouterj.nl>
  */
-class StopwatchTokenParser extends AbstractTokenParser
+final class StopwatchTokenParser extends AbstractTokenParser
 {
     protected $stopwatchIsAvailable;
 
@@ -31,10 +31,7 @@ class StopwatchTokenParser extends AbstractTokenParser
         $this->stopwatchIsAvailable = $stopwatchIsAvailable;
     }
 
-    /**
-     * @return Node
-     */
-    public function parse(Token $token)
+    public function parse(Token $token): Node
     {
         $lineno = $token->getLine();
         $stream = $this->parser->getStream();
@@ -55,15 +52,12 @@ class StopwatchTokenParser extends AbstractTokenParser
         return $body;
     }
 
-    public function decideStopwatchEnd(Token $token)
+    public function decideStopwatchEnd(Token $token): bool
     {
         return $token->test('endstopwatch');
     }
 
-    /**
-     * @return string
-     */
-    public function getTag()
+    public function getTag(): string
     {
         return 'stopwatch';
     }