+++ src/Symfony/Component/HttpKernel/Tests/KernelTest.php
@@ -21,6 +21,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Bundle\BundleInterface;
 use Symfony\Component\HttpKernel\DependencyInjection\ResettableServicePass;
 use Symfony\Component\HttpKernel\DependencyInjection\ServicesResetter;
+use Symfony\Component\HttpKernel\HttpKernel;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\HttpKernel\Tests\Fixtures\KernelForTest;
@@ -358,6 +359,9 @@ EOF;
         $this->assertEquals(__DIR__.'/Fixtures/Bundle1Bundle/foo.txt', $kernel->locateResource('@Bundle1Bundle/foo.txt'));
     }
 
+    /**
+     * @group legacy
+     */
     public function testLocateResourceIgnoresDirOnNonResource()
     {
         $kernel = $this->getKernel(['getBundle']);
@@ -373,6 +377,9 @@ EOF;
         );
     }
 
+    /**
+     * @group legacy
+     */
     public function testLocateResourceReturnsTheDirOneForResources()
     {
         $kernel = $this->getKernel(['getBundle']);
@@ -388,7 +395,10 @@ EOF;
         );
     }
 
-    public function testLocateResourceOnDirectories()
+    /**
+     * @group legacy
+     */
+    public function testLocateResourceOnDirectoriesWithOverwrite()
     {
         $kernel = $this->getKernel(['getBundle']);
         $kernel
@@ -405,7 +415,10 @@ EOF;
             __DIR__.'/Fixtures/Resources/FooBundle',
             $kernel->locateResource('@FooBundle/Resources', __DIR__.'/Fixtures/Resources')
         );
+    }
 
+    public function testLocateResourceOnDirectories()
+    {
         $kernel = $this->getKernel(['getBundle']);
         $kernel
             ->expects($this->exactly(2))
@@ -577,10 +590,8 @@ EOF;
 
     /**
      * Returns a mock for the BundleInterface.
-     *
-     * @return BundleInterface
      */
-    protected function getBundle($dir = null, $parent = null, $className = null, $bundleName = null)
+    protected function getBundle($dir = null, $parent = null, $className = null, $bundleName = null): BundleInterface
     {
         $bundle = $this
             ->getMockBuilder('Symfony\Component\HttpKernel\Bundle\BundleInterface')
@@ -620,10 +631,8 @@ EOF;
      *
      * @param array $methods Additional methods to mock (besides the abstract ones)
      * @param array $bundles Bundles to register
-     *
-     * @return Kernel
      */
-    protected function getKernel(array $methods = [], array $bundles = [])
+    protected function getKernel(array $methods = [], array $bundles = []): Kernel
     {
         $methods[] = 'registerBundles';
 
@@ -661,7 +670,7 @@ class TestKernel implements HttpKernelInterface
         $this->terminateCalled = true;
     }
 
-    public function handle(Request $request, int $type = self::MASTER_REQUEST, bool $catch = true)
+    public function handle(Request $request, int $type = self::MASTER_REQUEST, bool $catch = true): Response
     {
     }
 }
@@ -680,7 +689,7 @@ class CustomProjectDirKernel extends Kernel
         $this->httpKernel = $httpKernel;
     }
 
-    public function registerBundles()
+    public function registerBundles(): iterable
     {
         return [];
     }
@@ -689,7 +698,7 @@ class CustomProjectDirKernel extends Kernel
     {
     }
 
-    public function getProjectDir()
+    public function getProjectDir(): string
     {
         return __DIR__.'/Fixtures';
     }
@@ -701,7 +710,7 @@ class CustomProjectDirKernel extends Kernel
         }
     }
 
-    protected function getHttpKernel()
+    protected function getHttpKernel(): HttpKernelInterface
     {
         return $this->httpKernel;
     }